requirements.txt
absl-py==0.5.0
asn1crypto==0.23.0
astor==0.7.1
astroid==1.6.5
autopep8==1.4.2
awscli==1.14.29
awslogs==0.10.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bcrypt==3.1.3
boto3==1.4.7
botocore==1.7.17
certifi==2018.4.16
cffi==1.11.0
chardet==3.0.4
click==6.7
colorama==0.3.7
configparser==3.5.0
contextlib2==0.5.5
cryptography==2.0.3
docutils==0.14
enum34==1.1.6
Flask==1.0.2
funcsigs==1.0.2
futures==3.1.1
gast==0.2.0
gevent==1.3.2.post0
gevent-websocket==0.10.1
gitdb2==2.0.2
GitPython==2.1.6
greenlet==0.4.13
grpcio==1.15.0
idna==2.6
ipaddress==1.0.18
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
lxml==4.2.1
Markdown==3.0
MarkupSafe==1.0
mccabe==0.6.1
mitmproxy==0.11.1
mock==2.0.0
netlib==0.11.2
numpy==1.14.5
paramiko==2.3.1
passlib==1.7.1
pbr==4.2.0
Pillow==5.3.0
protobuf==3.6.1
psutil==5.4.5
pyaml==17.8.0
pyasn1==0.3.6
pycodestyle==2.4.0
pycparser==2.18
PyJWT==1.6.4
pylint==1.9.3
PyNaCl==1.1.2
pystache==0.5.4
pytesseract==0.2.5
python-dateutil==2.6.1
pytz==2018.5
PyYAML==3.12
raven==6.9.0
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.11
selenium==3.13.0
singledispatch==3.4.0.3
six==1.11.0
slack-cleaner==0.3.0
slacker==0.9.65
smmap2==2.0.3
sortedcontainers==2.0.3
splinter==0.8.0
tamper==0.24.5
TBB==0.1
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
tesseract==0.1.3
tornado==5.0.2
twilio==6.16.0
urllib3==1.23
urwid==2.0.1
websocket-client==0.48.0
Werkzeug==0.14.1
wrapt==1.10.11
